RandBats: Refactor routing of generators for min hardcoding
Note: Regeneration was incorrectly commited with CRLF before.
This commit also fixes that.
