+++ data/random-battles/afd/sets.json
