+++ data/random-battles/gen8/index.ts
