+++ data/random-battles/gen9/index.ts
@@ -3,7 +3,7 @@ import { Utils } from '../../../lib';
 import { PRNG, type PRNGSeed } from '../../../sim/prng';
 import { type RuleTable } from '../../../sim/dex-formats';
 import { Tags } from './../../tags';
-import { Teams } from '../../../sim/teams';
+import { Teams } from '../../../sim/index';
 
 export interface TeamData {
 	typeCount: { [k: string]: number };
