+++ data/random-battles/randomroulette/index.ts
@@ -1,4 +1,4 @@
-import RandomTeams from '../gen9/teams';
+import RandomTeams from '../gen9/index';
 
 export class RandomRandomRouletteTeams extends RandomTeams {}
